+++ packages/flutter_form_bloc/example/lib/forms/crud_form.dart
@@ -1,0 +1,93 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_form_bloc/flutter_form_bloc.dart';
+import 'package:flutter_form_bloc_example/forms/crud_form_bloc.dart';
+import 'package:flutter_form_bloc_example/widgets/widgets.dart';
+import 'package:form_bloc/form_bloc.dart';
+
+class CrudForm extends StatelessWidget {
+  final String name;
+
+  const CrudForm({Key key, this.name}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocProvider<CrudFormBloc>(
+      create: (context) => CrudFormBloc(name: name),
+      child: Builder(
+        builder: (context) {
+          final formBloc = BlocProvider.of<CrudFormBloc>(context);
+
+          return Scaffold(
+            appBar: AppBar(title: Text('CRUD')),
+            body: FormBlocListener<CrudFormBloc, String, String>(
+              onSubmitting: (context, state) => LoadingDialog.show(context),
+              onSuccess: (context, state) {
+                LoadingDialog.hide(context);
+                Notifications.showSnackBarWithSuccess(
+                    context, state.isEditing ? 'Updated!' : 'Created!');
+              },
+              onFailure: (context, state) {
+                LoadingDialog.hide(context);
+                Notifications.showSnackBarWithError(
+                    context,
+                    state.isEditing
+                        ? 'Error trying to create'
+                        : 'Error trying to update');
+              },
+              onDeleting: (context, state) => LoadingDialog.show(context),
+              onDeleteFailed: (context, state) {
+                LoadingDialog.hide(context);
+
+                Notifications.showSnackBarWithError(context, 'Delete Failed');
+              },
+              onDeleteSuccessful: (context, state) {
+                LoadingDialog.hide(context);
+                Navigator.of(context).pop();
+                Notifications.showSnackBarWithSuccess(
+                    context, 'Deleted Successful!');
+              },
+              child: ListView(
+                physics: ClampingScrollPhysics(),
+                children: <Widget>[
+                  TextFieldBlocBuilder(
+                    textFieldBloc: formBloc.nameField,
+                    keyboardType: TextInputType.emailAddress,
+                    decoration: InputDecoration(
+                      labelText: 'Name',
+                      prefixIcon: Icon(Icons.person),
+                    ),
+                  ),
+                  BlocBuilder<CrudFormBloc, FormBlocState>(
+                    builder: (context, state) {
+                      return Padding(
+                        padding: const EdgeInsets.all(8.0),
+                        child: RaisedButton(
+                          onPressed: formBloc.submit,
+                          child: Center(
+                              child:
+                                  Text(state.isEditing ? 'UPDATE' : 'CREATE')),
+                        ),
+                      );
+                    },
+                  ),
+                  BlocBuilder<CrudFormBloc, FormBlocState>(
+                    builder: (context, state) {
+                      return Padding(
+                        padding: const EdgeInsets.all(8.0),
+                        child: RaisedButton(
+                          onPressed: state.isEditing ? formBloc.delete : null,
+                          child: Center(child: Text('DELETE')),
+                        ),
+                      );
+                    },
+                  ),
+                ],
+              ),
+            ),
+          );
+        },
+      ),
+    );
+  }
+}